pyproject.toml
[tool.poetry]
name = "avaris"
version = "0.2.0"
description = "Task execution engine for data management and monitoring"
authors = ["dennis <denngohis@gmail.com>"]
license = "Apache License 2.0"
readme = "README.md"
packages = [{ include = "avaris", from = "src" }]

[tool.poetry.scripts]
avaris = "avaris.cli:avaris"


[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.6.3"
python-dotenv = { extras = ["cli"], version = "^1.0.1" }
click = "^8.1.7"
celery = "^5.3.6"
apscheduler = "^3.10.4"
requests = "^2.31.0"
croniter = "^2.0.2"
aiofiles = "^23.2.1"
aiohttp = "^3.9.3"
pyyaml = "^6.0.1"
flask = "^3.0.2"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.28"
databases = "^0.9.0"
aiosqlite = "^0.20.0"
pytz = "^2024.1"
httpx = "^0.27.0"
colorlog = "^6.8.2"
beautifulsoup4 = "^4.12.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
pytest-httpx = "^0.30.0"
flake8 = "^7.0.0"
pytest-mock = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
